+++ CHANGELOG.md
@@ -9,7 +9,7 @@
   - DepositProvider#deposit -> #startDeposit
 - Transfers: Fix fetchFinalFee parameters
 - Transfers: Add `DepositProvider#validateFields` to validate deposit fields
-- Transfers: Add `TransferProvider#getAsset` to fetch single assets
+- Transfers: Add `TransferProvider#getAssetInfo` to fetch single assets
 
 ## [v0.0.6-rc.17](https://github.com/stellar/js-stellar-wallets/compare/v0.0.3-rc1...v0.0.6-rc.17)